- Shamelessly stolen payload from https://github.com/B-Dem/PPPwnUI/blob/main/PPPwn/goldhen/1100/ - Auto restart if PPPwn fails - Automatically shut down the PC if PPPwn takes more then 5 minutes

+++ custom/genapkovl-pppwn.sh
@@ -91,32 +91,43 @@ EOF
 
 makefile root:root 0755 "$tmp"/etc/setup.sh <<EOF
 #!/bin/sh
+
 tar -xzf /etc/pppwn.tar.gz -C /root/
 chmod +x /root/pppwnlive/pppwn
+
 # Find the first available ethernet interface
 ETH_IF=\$(ip -o link show | awk -F': ' '\$2 ~ /^eth|^en/ {print \$2; exit}')
+
 # Enable the ethernet interface
 ip link set dev \$ETH_IF up
 
 if [ -n "\$ETH_IF" ]; then
     # Run pppwn command with the detected interface
     cd /root/pppwnlive
-    if ./pppwn -i "\$ETH_IF" --fw 1100 --stage1 stage1.bin --stage2 stage2.bin -a; then
-        echo "PPPwn executed successfully. Shutting down..."
-        sleep 3
-        poweroff
-    else
-        echo "PPPwn failed to execute successfully."
-        echo "Press 'q' to exit without shutting down, or any other key to shutdown..."
-        read -n 1 -s key
-        if [ "\$key" = "q" ]; then
-            echo "Exiting without shutdown."
-            exit 1
-        else
-            echo "Shutting down..."
+    while true; do
+        if timeout 300 ./pppwn -i "\$ETH_IF" --fw 1100 --stage1 stage1.bin --stage2 stage2.bin -a; then
+            echo "PPPwn executed successfully. Shutting down..."
+            sleep 3
             poweroff
+        else
+            if [ \$? -eq 124 ]; then
+                echo "PPPwn timed out after 5 minutes. Shutting down..."
+                sleep 3
+                poweroff
+            else
+                echo "PPPwn failed to execute successfully."
+                echo "Press 'q' to exit without shutting down, or wait 3 seconds for automatic restart..."
+                if timeout 3 read -n 1 -s key; then
+                    if [ "\$key" = "q" ]; then
+                        echo "Exiting without shutdown."
+                        exit 1
+                    fi
+                else
+                    echo "No input received. Restarting script..."
+                fi
+            fi
         fi
-    fi
+    done
 else
     echo "No ethernet interface found. Please check your connection."
     echo "Press any key to shutdown..."
@@ -124,6 +135,7 @@ else
     poweroff
 fi
 EOF
+
 # Use /root/.profile to automatically run /etc/setup.sh
 makefile root:root 0644 "$tmp"/etc/.profile <<EOF
 /etc/setup.sh
